migrate usecases layout
