+++ src/layouts/index.ts
@@ -3,4 +3,3 @@ export * from "./Docs"
 export * from "./md"
 export * from "./Static"
 export * from "./Tutorial"
-export * from "./UseCases"
