+++ src/layouts/md/index.ts
@@ -2,3 +2,4 @@ export * from "./Roadmap"
 export * from "./Staking"
 export * from "./Translatathon"
 export * from "./Upgrade"
+export * from "./UseCases"
